backend/agents/infra_agent/requirements.txt
google-generativeai==0.3.2
flask==2.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
pyyaml==6.0.1